Added trinucleotides field to the signature to be used for indel without signature.
Indels simulated as substitutions may use the signature or not.
If not, then it needs a probability.
Typically it should be 1/64 if all the trinucleotides are present.
Otherwise, we should use the inverse of the number of trinucleotides seen in all the regions.
We use the value obtained from the count in the correction of the signature if it is done.
If not, the default value 1/64 is used.